LG-666 Refactor AccountResetService
**Why**: To make it more consistent with the other classes under the
AccountReset namespace.
