+++ lib/tasks/account_reset.rake
@@ -1,6 +1,6 @@
 namespace :account_reset do
   desc 'Send Notifications'
   task send_notifications: :environment do
-    AccountResetService.grant_tokens_and_send_notifications
+    AccountReset::GrantRequestsAndSendEmails.new.call
   end
 end
