+++ spec/features/account_reset/cancel_request_spec.rb
@@ -35,7 +35,7 @@ describe 'Account Reset Request: Cancellation' do
       reset_email
 
       Timecop.travel(Time.zone.now + 2.days) do
-        AccountResetService.grant_tokens_and_send_notifications
+        AccountReset::GrantRequestsAndSendEmails.new.call
         open_last_email
         click_email_link_matching(/cancel\?token/)
         click_button t('account_reset.cancel_request.cancel_button')
